requirements.txt
# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
ruff>=0.1.0
